Remove no longer needed celery tasks
