+++ temba/flows/legacy/engine.py
@@ -10,7 +10,7 @@ from django.db.models import Q, QuerySet
 from django.utils import timezone
 
 from temba.contacts.models import Contact, ContactGroup
-from temba.msgs.models import DELIVERED, FAILED, FLOW, OUTGOING, PENDING, Msg
+from temba.msgs.models import DELIVERED, FAILED, FLOW, INBOX, INCOMING, OUTGOING, PENDING, Msg
 from temba.utils import json, prepped_request_to_str
 from temba.utils.http import http_headers
 
@@ -126,12 +126,12 @@ def flow_start(  # pragma: no cover
         .exclude(id__in=ancestor_ids)
         .exclude(flow__is_system=True)
     )
-    FlowRun.bulk_exit(active_runs, FlowRun.EXIT_TYPE_INTERRUPTED)
+    bulk_exit(active_runs, FlowRun.EXIT_TYPE_INTERRUPTED)
 
     # if we are interrupting parent flow runs, mark them as completed
     if ancestor_ids and interrupt:
         ancestor_runs = FlowRun.objects.filter(id__in=ancestor_ids)
-        FlowRun.bulk_exit(ancestor_runs, FlowRun.EXIT_TYPE_COMPLETED)
+        bulk_exit(ancestor_runs, FlowRun.EXIT_TYPE_COMPLETED)
 
     if not all_contact_ids:
         return []
@@ -233,15 +233,15 @@ def _flow_start(flow, contact_ids, started_flows=None, start_msg=None, extra=Non
             if entry_actions:
                 run_msgs += entry_actions.execute_actions(run, start_msg, started_flows_by_contact)
 
-                add_step(run, entry_actions, run_msgs, arrived_on=arrived_on)
+                _add_step(run, entry_actions, run_msgs, arrived_on=arrived_on)
 
                 # and onto the destination
                 if entry_actions.destination:
-                    destination = Flow.get_node(
+                    destination = get_node(
                         entry_actions.flow, entry_actions.destination, entry_actions.destination_type
                     )
 
-                    add_step(run, destination, exit_uuid=entry_actions.exit_uuid)
+                    _add_step(run, destination, exit_uuid=entry_actions.exit_uuid)
 
                     msg = Msg(org=flow.org, contact=contact, text="", id=0)
                     handled, step_msgs = _handle_destination(
@@ -253,7 +253,7 @@ def _flow_start(flow, contact_ids, started_flows=None, start_msg=None, extra=Non
                     run.set_completed(exit_uuid=None)
 
             elif entry_rules:
-                add_step(run, entry_rules, run_msgs, arrived_on=arrived_on)
+                _add_step(run, entry_rules, run_msgs, arrived_on=arrived_on)
 
                 # if we have a start message, go and handle the rule
                 if start_msg:
@@ -302,7 +302,7 @@ def _flow_start(flow, contact_ids, started_flows=None, start_msg=None, extra=Non
     return runs
 
 
-def add_step(run, node, msgs=(), exit_uuid=None, arrived_on=None):
+def _add_step(run, node, msgs=(), exit_uuid=None, arrived_on=None):
     """
     Adds a new step to the given run
     """
@@ -332,7 +332,7 @@ def add_step(run, node, msgs=(), exit_uuid=None, arrived_on=None):
     update_fields = ["path", "current_node_uuid"]
 
     if msgs:
-        run.add_messages(msgs, do_save=False)
+        _add_messages(run, msgs, do_save=False)
         update_fields += ["responded", "events"]
 
     run.current_node_uuid = run.path[-1][FlowRun.PATH_NODE_UUID]
@@ -366,7 +366,7 @@ def find_and_handle(
                 return False, []
 
         last_step = run.path[-1]
-        destination = Flow.get_node(flow, last_step[FlowRun.PATH_NODE_UUID], Flow.NODE_TYPE_RULESET)
+        destination = get_node(flow, last_step[FlowRun.PATH_NODE_UUID], Flow.NODE_TYPE_RULESET)
 
         # this node doesn't exist anymore, mark it as left so they leave the flow
         if not destination:  # pragma: no cover
@@ -402,7 +402,7 @@ def _handle_destination(
     resume_parent_run=False,
     continue_parent=True,
 ):
-    from temba.flows.models import Flow, FlowRun, FlowException
+    from temba.flows.models import Flow, FlowException
 
     def add_to_path(path, uuid):
         if uuid in path:
@@ -427,7 +427,7 @@ def _handle_destination(
                 should_pause = True
 
             if user_input or not should_pause:
-                result = Flow.handle_ruleset(destination, run, msg, started_flows, resume_parent_run)
+                result = _handle_ruleset(destination, run, msg, started_flows, resume_parent_run)
                 add_to_path(path, destination.uuid)
 
                 # add any messages generated by this ruleset
@@ -441,7 +441,7 @@ def _handle_destination(
                 path = []
 
         elif destination.get_step_type() == Flow.NODE_TYPE_ACTIONSET:
-            result = Flow.handle_actionset(destination, run, msg, started_flows)
+            result = _handle_actionset(destination, run, msg, started_flows)
             add_to_path(path, destination.uuid)
 
             # add any generated messages to be sent at once
@@ -462,7 +462,7 @@ def _handle_destination(
 
     # if we have a parent to continue, do so
     if getattr(run, "continue_parent", False) and continue_parent:
-        msgs += FlowRun.continue_parent_flow_run(run, trigger_send=False, continue_parent=True)
+        msgs += _continue_parent_run(run, trigger_send=False, continue_parent=True)
 
     # send any messages generated
     if msgs and trigger_send:
@@ -471,6 +471,255 @@ def _handle_destination(
         run.flow.org.trigger_send(msgs)
 
     return handled, msgs
+
+
+def _handle_actionset(actionset, run, msg, started_flows):
+    # not found, escape out, but we still handled this message, user is now out of the flow
+    if not actionset:  # pragma: no cover
+        run.set_completed(exit_uuid=None)
+        return dict(handled=True, destination=None, destination_type=None)
+
+    # actually execute all the actions in our actionset
+    msgs = actionset.execute_actions(run, msg, started_flows)
+    _add_messages(run, [m for m in msgs if not getattr(m, "from_other_run", False)])
+
+    # and onto the destination
+    destination = get_node(actionset.flow, actionset.destination, actionset.destination_type)
+    if destination:
+        _add_step(run, destination, exit_uuid=actionset.exit_uuid)
+    else:
+        run.set_completed(exit_uuid=actionset.exit_uuid)
+
+    return dict(handled=True, destination=destination, msgs=msgs)
+
+
+def _handle_ruleset(ruleset, run, msg_in, started_flows, resume_parent_run=False):
+    from temba.flows.models import RuleSet, Flow, FlowRun
+
+    msgs_out = []
+    result_input = str(msg_in)
+
+    if ruleset.ruleset_type == RuleSet.TYPE_SUBFLOW:
+        if not resume_parent_run:
+            flow_uuid = ruleset.config.get("flow").get("uuid")
+            flow = Flow.objects.filter(org=run.org, uuid=flow_uuid).first()
+            flow.org = run.org
+            message_context = run.flow.build_expressions_context(run.contact, msg_in, run=run)
+
+            # our extra will be the current flow variables
+            extra = message_context.get("extra", {})
+            extra["flow"] = message_context.get("flow", {})
+
+            if msg_in.id:
+                _add_messages(run, [msg_in])
+                run.update_expiration(timezone.now())
+
+            if flow:
+                child_runs = flow_start(
+                    flow,
+                    [],
+                    [run.contact],
+                    started_flows=started_flows,
+                    restart_participants=True,
+                    extra=extra,
+                    parent_run=run,
+                    interrupt=False,
+                )
+
+                child_run = child_runs[0] if child_runs else None
+
+                if child_run:
+                    msgs_out += child_run.start_msgs
+                    continue_parent = getattr(child_run, "continue_parent", False)
+                else:  # pragma: no cover
+                    continue_parent = False
+
+                # it's possible that one of our children interrupted us with a start flow action
+                run.refresh_from_db(fields=("is_active",))
+                if continue_parent and run.is_active:
+                    started_flows.remove(flow.id)
+
+                    run.child_context = child_run.build_expressions_context(contact_context=str(run.contact.uuid))
+                    run.save(update_fields=("child_context",))
+                else:
+                    return dict(handled=True, destination=None, destination_type=None, msgs=msgs_out)
+
+        else:
+            child_run = FlowRun.objects.filter(parent=run, contact=run.contact).order_by("created_on").last()
+            run.child_context = child_run.build_expressions_context(contact_context=str(run.contact.uuid))
+            run.save(update_fields=("child_context",))
+
+    # find a matching rule
+    result_rule, result_value, result_input = ruleset.find_matching_rule(run, msg_in)
+
+    flow = ruleset.flow
+
+    # add the message to our step
+    if msg_in.id:
+        _add_messages(run, [msg_in])
+        run.update_expiration(timezone.now())
+
+    if ruleset.ruleset_type in RuleSet.TYPE_MEDIA and msg_in.attachments:  # pragma: no cover
+        # store the media path as the value
+        result_value = msg_in.attachments[0].split(":", 1)[1]
+
+    ruleset.save_run_value(run, result_rule, result_value, result_input, org=flow.org)
+
+    # no destination for our rule?  we are done, though we did handle this message, user is now out of the flow
+    if not result_rule.destination:
+        run.set_completed(exit_uuid=result_rule.uuid)
+        return dict(handled=True, destination=None, destination_type=None, msgs=msgs_out)
+
+    # Create the step for our destination
+    destination = get_node(flow, result_rule.destination, result_rule.destination_type)
+    if destination:
+        _add_step(run, destination, exit_uuid=result_rule.uuid)
+
+    return dict(handled=True, destination=destination, msgs=msgs_out)
+
+
+def _add_messages(run, msgs, do_save=True):
+    """
+    Associates the given messages with a run
+    """
+    from temba.flows.models import FlowRun, Events
+
+    if run.events is None:
+        run.events = []
+
+    # find the path step these messages belong to
+    path_step = run.path[-1]
+
+    existing_msg_uuids = set()
+    for e in run.get_msg_events():
+        msg_uuid = e["msg"].get("uuid")
+        if msg_uuid:
+            existing_msg_uuids.add(msg_uuid)
+
+    needs_update = False
+
+    def serialize_message(msg):
+        serialized = {"uuid": str(msg.uuid), "text": msg.text}
+
+        if msg.contact_urn_id:
+            serialized["urn"] = msg.contact_urn.urn
+        if msg.channel_id:
+            serialized["channel"] = {"uuid": str(msg.channel.uuid), "name": msg.channel.name or ""}
+        if msg.attachments:
+            serialized["attachments"] = msg.attachments
+
+        return serialized
+
+    for msg in msgs:
+        # or messages which have already been attached to this run
+        if str(msg.uuid) in existing_msg_uuids:
+            continue
+
+        run.events.append(
+            {
+                FlowRun.EVENT_TYPE: Events.msg_received.name if msg.direction == INCOMING else Events.msg_created.name,
+                FlowRun.EVENT_CREATED_ON: msg.created_on.isoformat(),
+                FlowRun.EVENT_STEP_UUID: path_step.get(FlowRun.PATH_STEP_UUID),
+                "msg": serialize_message(msg),
+            }
+        )
+
+        existing_msg_uuids.add(str(msg.uuid))
+        needs_update = True
+
+        # incoming non-IVR messages won't have a type yet so update that
+        if not msg.msg_type or msg.msg_type == INBOX:
+            msg.msg_type = FLOW
+            msg.save(update_fields=["msg_type"])
+
+        # if message is from contact, mark run as responded
+        if msg.direction == INCOMING:
+            if not run.responded:
+                run.responded = True
+
+    if needs_update and do_save:
+        run.save(update_fields=("responded", "events"))
+
+
+def bulk_exit(runs, exit_type):
+    """
+    Exits (expires, interrupts) runs in bulk
+    """
+
+    from temba.flows.models import FlowRun, FlowSession
+
+    now = timezone.now()
+
+    run_ids = list(runs[:5000].values_list("id", flat=True))
+    runs = FlowRun.objects.filter(id__in=run_ids)
+    runs.update(
+        is_active=False, exited_on=now, exit_type=exit_type, modified_on=now, child_context=None, parent_context=None
+    )
+
+    for run in runs:
+        if run.parent and run.parent.is_active and run.parent.flow.is_active and not run.parent.flow.is_archived:
+            _continue_parent_run(run)
+
+    # mark session as completed if this is an interruption
+    if exit_type == FlowRun.EXIT_TYPE_INTERRUPTED:
+        (
+            FlowSession.objects.filter(id__in=runs.exclude(session=None).values_list("session_id", flat=True))
+            .filter(status=FlowSession.STATUS_WAITING)
+            .update(status=FlowSession.STATUS_INTERRUPTED, ended_on=now)
+        )
+
+
+def _continue_parent_run(run, trigger_send=True, continue_parent=True):
+    from temba.flows.models import FlowRun, RuleSet
+
+    # TODO: Remove this in favor of responded on session
+    if run.responded and not run.parent.responded:
+        run.parent.responded = True
+        run.parent.save(update_fields=["responded"])
+
+    # if our child was interrupted, so shall we be
+    if run.exit_type == FlowRun.EXIT_TYPE_INTERRUPTED and run.contact.id == run.parent.contact_id:
+        bulk_exit(FlowRun.objects.filter(id=run.parent_id), FlowRun.EXIT_TYPE_INTERRUPTED)
+        return
+
+    last_step = run.parent.path[-1]
+    ruleset = (
+        RuleSet.objects.filter(
+            uuid=last_step[FlowRun.PATH_NODE_UUID], ruleset_type=RuleSet.TYPE_SUBFLOW, flow__org=run.org
+        )
+        .exclude(flow=None)
+        .first()
+    )
+
+    # can't resume from a ruleset that no longer exists
+    if not ruleset:
+        return []
+
+    # use the last incoming message on this run
+    msg = run.get_last_msg(direction=INCOMING)
+
+    # if we are routing back to the parent before a msg was sent, we need a placeholder
+    if not msg:
+        msg = Msg()
+        msg.id = 0
+        msg.text = ""
+        msg.org = run.org
+        msg.contact = run.contact
+
+    run.parent.child_context = run.build_expressions_context(contact_context=str(run.contact.uuid))
+    run.parent.save(update_fields=("child_context",))
+
+    # finally, trigger our parent flow
+    (handled, msgs) = find_and_handle(
+        msg,
+        user_input=False,
+        started_flows=[run.flow, run.parent.flow],
+        resume_parent_run=True,
+        trigger_send=trigger_send,
+        continue_parent=continue_parent,
+    )
+
+    return msgs
 
 
 def call_webhook(run, webhook_url, ruleset, msg, action="POST", resthook=None, headers=None):
@@ -605,3 +854,19 @@ def call_webhook(run, webhook_url, ruleset, msg, action="POST", resthook=None, h
         )
 
     return result
+
+
+def get_node(flow, uuid, destination_type):
+    from temba.flows.models import Flow, ActionSet, RuleSet
+
+    if not uuid or not destination_type:
+        return None
+
+    if destination_type == Flow.NODE_TYPE_RULESET:
+        node = RuleSet.get(flow, uuid)
+    else:
+        node = ActionSet.get(flow, uuid)
+
+    if node:
+        node.flow = flow
+    return node
