+++ temba/flows/tasks.py
@@ -46,16 +46,14 @@ def update_run_expirations_task(flow_id):
             run.update_expiration(last_arrived_on)
 
 
-@task(track_started=True, name="continue_parent_flows")  # pragma: no cover
-def continue_parent_flows(run_ids):
-    runs = FlowRun.objects.filter(pk__in=run_ids)
-    FlowRun.continue_parent_flow_runs(runs)
-
-
 @task(track_started=True, name="interrupt_flow_runs_task")
 def interrupt_flow_runs_task(flow_id):
+    from temba.flows import legacy
+
     runs = FlowRun.objects.filter(is_active=True, exit_type=None, flow_id=flow_id)
-    FlowRun.bulk_exit(runs, FlowRun.EXIT_TYPE_INTERRUPTED)
+
+    # TODO this should be queuing a mailroom task
+    legacy.bulk_exit(runs, FlowRun.EXIT_TYPE_INTERRUPTED)
 
 
 @task(track_started=True, name="export_flow_results_task")
